Home

ProgressLogging.jl

ProgressLogging.progress

ProgressLogging.@progress

ProgressLogging.@progress — Macro.

@progress [name="", threshold=0.005] for i = ..., j = ..., ...
@progress [name="", threshold=0.005] x = [... for i = ..., j = ..., ...]

Show a progress meter named name for the given loop or array comprehension if possible. Update frequency is limited by threshold (one update per 0.5% of progress by default).

source

ProgressLogging.progress — Method.

progress(f::Function; name = "")

Evaluates f with id as its argument and makes sure to destroy the progress bar afterwards. To update the progress bar in f you can call a logging statement like @info or even just @logmsg with _id=id and progress as arguments.

progress can take either of the following values:

0 <= progress < 1: create or update progress bar

progress == nothing || progress = NaN: set progress bar to indeterminate progress

progress >= 1 || progress == "done": destroy progress bar

The logging message (e.g. "foo" in @info "foo") will be used as the progress bar's name.

Log level must be higher or equal to LogLevel(-1).

ProgressLogging.progress() do id
    for i = 1:10
        sleep(0.5)
        @info "iterating" progress=i/10 _id=id
    end
end

source